cleaning up type inference
